pyproject.toml
[tool.poetry]
name = "kafka-images-streaming"
version = "0.1.0"
description = ""
authors = ["Tom Jansen <tomjansen25@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10, <3.12"
numpy = "^2.1.1"
kafka-python = "^2.0.2"
pymongo = "^4.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
